Move DiceTotal to its own mod
